+++ conf/jsdocConfig.js
@@ -1,3 +1,5 @@
+/* eslint-env node */
+
 /**
  * @module conf/jsdocConfig
  */
@@ -14,8 +16,8 @@ var extend = require('../lib/deepExtend'),
  * [ink-docstrap](https://github.com/terryweiss/docstrap),
  * `lenient` will set set to true, and docs will go in `src/jsdoc`.
  *
- * @param {external:Grunt} grunt
- * @returns {Object} `jsdoc` Grunt configuration object
+ * @param {IGrunt} grunt
+ * @returns {Object} configuration for `jsdoc` task
  */
 module.exports = function (grunt) {
   var defaultJsdocPath = path.resolve(require.resolve('grunt-jsdoc'), '..', 'jsdoc', 'jsdoc'),
@@ -28,11 +30,11 @@ module.exports = function (grunt) {
           lenient: true,
           recurse: true,
           template: defaultTemplatePath,
-          destination: 'src/jsdoc'
+          destination: 'build/src/jsdoc'
         },
-        dist : {
+        dist: {
           src: src,
-          jsdoc: defaultJsdocPath,
+          jsdoc: defaultJsdocPath
         }
       };