+++ tasks/default.js
@@ -1,3 +1,5 @@
+/* eslint-env node */
+
 /** @module */
 
 'use strict';
@@ -5,7 +7,7 @@
 /**
  * The `default` task prints out usage information.
  *
- * @param {Grunt} grunt
+ * @param {IGrunt} grunt
  */
 module.exports = function (grunt) {
   grunt.registerTask('default', [ 'usage' ]);