+++ tasks/karma.js
@@ -1,17 +1,19 @@
+/* eslint-env node */
+
 /** @module */
 
 'use strict';
 
 var parseOptions = require('../lib/parseOptions'),
-    niagaraConfig = require('../conf/niagaraConfig'),
-    karmaConfig = require('../conf/karmaConfig'),
-    runKarma = require('../lib/runKarma');
+  niagaraConfig = require('../conf/niagaraConfig'),
+  karmaConfig = require('../conf/karmaConfig'),
+  runKarma = require('../lib/runKarma');
 
 /**
  * The `karma` task runs unit tests using Karma. `karma:ci` also generates
  * code coverage reports.
  *
- * @param {Grunt} grunt
+ * @param {IGrunt} grunt
  */
 module.exports = function (grunt) {
   if (!grunt.config('karma')) {
@@ -22,9 +24,9 @@ module.exports = function (grunt) {
 
   grunt.config('karma', karmaConfig(grunt));
 
-  grunt.registerTask('karma', 'Run tests in Karma', function (type) {
-    var opts = parseOptions(grunt),
-        done = this.async();
+  grunt.registerTask('runKarma', 'Run tests in Karma', function (type) {
+    const opts = parseOptions(grunt);
+    const done = this.async();
 
     runKarma(grunt, type, this.flags.run, function (err) {
       if (err) {
@@ -35,5 +37,21 @@ module.exports = function (grunt) {
 
       return done(!err);
     });
+  });
+
+  grunt.registerTask('karma', function () {
+    const args = Object.keys(this.flags).join(':');
+    let tasks = [ 'runKarma' + (args ? ':' + args : '') ];
+
+    if (grunt.config('babel') && !args) {
+      const opts = parseOptions(grunt);
+      if (String(opts.coverage) === 'true') {
+        tasks = [ 'babel:coverage', 'babel:spec', 'copy:karma' ].concat(tasks);
+      } else {
+        tasks = [ 'babel:watch', 'babel:spec', 'copy:karma' ].concat(tasks);
+      }
+    }
+
+    grunt.task.run(tasks);
   });
 };