+++ tasks/usage.js
@@ -1,3 +1,5 @@
+/* eslint-env node */
+
 /** @module */
 
 'use strict';
@@ -8,16 +10,15 @@ var path = require('path'),
 /**
  * The `usage` task prints out information on how to use `grunt-niagara`.
  *
- * @param {Grunt} grunt
+ * @param {IGrunt} grunt
  */
 module.exports = function (grunt) {
   grunt.registerTask('usage', 'Prints out usage information', function () {
     var fileName = path.resolve(__dirname, '../info/usage.hbs');
     grunt.log.write(templateAndPrint(fileName, {
-      jshint: grunt.config('jshint'),
+      eslint: grunt.config('eslint'),
       karma: grunt.config('karma'),
       jsdoc: grunt.config('jsdoc'),
-      plato: grunt.config('plato'),
       watch: grunt.config('watch')
     }));
   });